Implement ?call query arg, replace nanoajax with fetch polyfill, refactor form flow
