+++ package.json
@@ -37,10 +37,11 @@
     "classlist-polyfill": "^1.2.0",
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
-    "nanoajax": "^0.4.3",
     "native-promise-only": "^0.8.1",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
-    "react-transition-group": "^1.1.3"
+    "react-transition-group": "^1.1.3",
+    "url-polyfill": "^1.0.6",
+    "whatwg-fetch": "^2.0.3"
   }
 }
