+++ ts/url-polyfill.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="../typings/index.d.ts" />
+
+declare module 'url-polyfill' {
+	export var URL: any;
+}
