+++ typings/index.d.ts
@@ -2,6 +2,5 @@
 /// <reference path="globals/jquery/index.d.ts" />
 /// <reference path="globals/node/index.d.ts" />
 /// <reference path="modules/lodash/index.d.ts" />
-/// <reference path="modules/nanoajax/index.d.ts" />
 /// <reference path="modules/react-dom/index.d.ts" />
 /// <reference path="modules/react/index.d.ts" />
