--- typings/modules/nanoajax/index.d.ts
@@ -1,23 +1,0 @@
-// Generated by typings
-// Source: https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/337587de8c13868283993bfacdcdd1a0f3291e7f/nanoajax/index.d.ts
-declare module 'nanoajax' {
-// Type definitions for nanoajax v0.2.4
-// Project: https://github.com/yanatan16/nanoajax
-// Definitions by: Nathan Cahill <https://github.com/nathancahill/>
-// Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-
-
-interface RequestParameters {
-    url: string;
-    headers?: { [key: string]: string; };
-    body?: string | FormData;
-    method?: string;
-    cors?: boolean;
-}
-
-interface Callback {
-    (statusCode: number, response: string, request: XMLHttpRequest): any
-}
-
-export function ajax(params: RequestParameters, callback: Callback): XMLHttpRequest
-}
